LayoutTests/imported/w3c/web-platform-tests/webrtc/protocol/crypto-suite.https-expected.txt


FAIL tlsVersion is acceptable on data-only promise_test: Unhandled rejection with value: object "TypeError: undefined is not an object (evaluating 'pc1.sctp.transport')"
PASS tlsVersion is acceptable on video-only
FAIL dtlsCipher is acceptable on data-only promise_test: Unhandled rejection with value: object "TypeError: undefined is not an object (evaluating 'pc1.sctp.transport')"
FAIL dtlsCipher is acceptable on video-only assert_true: expected true got false
FAIL srtpCipher is acceptable on data-only promise_test: Unhandled rejection with value: object "TypeError: undefined is not an object (evaluating 'pc1.sctp.transport')"
PASS srtpCipher is acceptable on video-only
FAIL tlsGroup is acceptable on data-only promise_test: Unhandled rejection with value: object "TypeError: undefined is not an object (evaluating 'pc1.sctp.transport')"
FAIL tlsGroup is acceptable on video-only assert_true: Value present: expected true got false
